test: improve test cases

+++ packages/create-brisa/create-brisa.test.ts
@@ -16,7 +16,9 @@ const EXPECTED_INNER_FILES = [
 
 describe('create-brisa', () => {
   afterEach(async () => {
-    await $`rm -rf out`;
+    const files = fs.readdirSync('.');
+    if (files.includes('out')) await $`rm -rf out`;
+    if (files.some((f) => f.startsWith('with-'))) await $`rm -rf with-*`;
   });
 
   it("should create brisa correctly with the name of the project as 'out'", async () => {
@@ -49,15 +51,18 @@ describe('create-brisa', () => {
   });
 
   it('should "bun create brisa --example" list the examples', async () => {
-    $`mkdir out && cd out`;
     const res = await $`echo 0 | bun run ${CREATE_BRISA_PATH} --example`;
-    $`cd .. && rm -rf out`;
     const out = res.stdout.toString();
     expect(res.exitCode).toBe(0);
     expect(out).toContain(
       'Choose an example:\n\t0. Exit\n\t1. with-api-routes\n\t2. with-external-web-component\n\t3. with-i18n\n\t4. with-middleware\n\t5. with-sqlite\n\t6. with-tailwind\nEnter the number of the example: ',
     );
     expect(out).toContain('Bye!');
+  });
+
+  it('should "bun create brisa --example" list the examples', async () => {
+    await $`echo 1 | bun run ${CREATE_BRISA_PATH} --example`;
+    expect(fs.existsSync(join('with-api-routes', 'src', 'api'))).toBeTrue();
   });
 
   it('should "bun create brisa --example with-api-routes" copy the example correctly', async () => {
